+++ datatig/cli.py
@@ -11,18 +11,27 @@ def main():
 
     build_parser = subparsers.add_parser("build")
     build_parser.add_argument("source")
-    build_parser.add_argument("--staticsite", help="Location of Static Site")
+    build_parser.add_argument(
+        "--staticsiteoutput", help="Location of Static Site Output"
+    )
+    build_parser.add_argument("--sqliteoutput", help="Location of SQLite file Output")
 
     args = parser.parse_args()
 
     if args.subparser_name == "build":
 
-        if not args.staticsite:
+        staticsite_output = args.staticsiteoutput
+        sqlite_output = args.sqliteoutput
+
+        if not staticsite_output and not sqlite_output:
             print("You must specify one of the build options when running build.")
             exit(-1)
 
         datatig.process.go(
-            args.source, os.path.join(args.source, "datatig.json"), args.staticsite
+            args.source,
+            os.path.join(args.source, "datatig.json"),
+            staticsite_output=staticsite_output,
+            sqlite_output=sqlite_output,
         )