+++ tests/test_jsonreader.py
@@ -1,6 +1,8 @@
 import json
 import os
+import sqlite3
 import tempfile
+from contextlib import closing
 
 import datatig.process
 
@@ -13,9 +15,11 @@ def test_json_site():
     with tempfile.TemporaryDirectory() as staticsite_dir:
         # Process!
         datatig.process.go(
-            source_dir, os.path.join(source_dir, "datatig.json"), staticsite_dir
+            source_dir,
+            os.path.join(source_dir, "datatig.json"),
+            staticsite_output=staticsite_dir,
         )
-        # Test
+        # Test Static Site - JSON files
         with open(
             os.path.join(staticsite_dir, "type", "datas", "record", "1", "data.json")
         ) as fp:
@@ -26,3 +30,12 @@ def test_json_site():
         ) as fp:
             two_json = json.load(fp)
             assert {"title": "Two"} == two_json
+        # Test database
+        with closing(
+            sqlite3.connect(os.path.join(staticsite_dir, "database.sqlite"))
+        ) as connection:
+            connection.row_factory = sqlite3.Row
+            with closing(connection.cursor()) as cur:
+                cur.execute("SELECT * FROM type")
+                type = cur.fetchone()
+                assert "datas" == type["id"]